refactor pipeline test scripts
