+++ apex/transformer/testing/commons.py
@@ -21,6 +21,7 @@ import torch
 import torch.nn as nn
 
 from apex import transformer
+from apex.transformer.pipeline_parallel.utils import average_losses_across_data_parallel_group
 from apex.transformer.testing import global_vars
 
 
@@ -61,6 +62,27 @@ class MyModel(nn.Module):
 
 def model_provider_func(hidden_size, pre_process, post_process) -> MyModel:
     return MyModel(hidden_size, pre_process, post_process)
+
+
+def process_batch(batch):
+    if isinstance(batch, list):
+        x = batch[0]
+    else:
+        x = batch
+    return x
+
+
+def fwd_step_func(batch, model):
+    x = process_batch(batch)
+    y = model(x)
+
+    # note (mkozuki): I don't think this function is nice but I do think this is enough for now
+    # just to check the sanity of ported pipeline functions.
+    def loss_func(x):
+        loss = torch.sum(x)
+        averaged_loss = average_losses_across_data_parallel_group([loss])
+        return loss, {'avg': averaged_loss}
+    return y, loss_func
 
 
 class IdentityLayer(torch.nn.Module):
